"resource_installer.py" should become packaged
1) Addressing @suvayu's comment https://github.com/tudat-team/tudatpy/issues/875#issuecomment-4927912026:
   Modules under /src/tudatpy will be installed while a module as a top level script will not. Now e.g. this will work
   "python -m tudatpy.resource_installer --mode missing   --dest ~/Test   --cache-dir /tmp/tudatpy_resources_test --hash-file resource_hashes.json"
2) Added a new module "src/tudatpy/resource_data_registry.py" to ensure that "src/tudatpy/resource_installer.py" only does installation and not the control of the (contents of) the catalog: this has been moved to "src/tudatpy/resource_data_registry.py".
