+++ cmake/tudatpy/PythonExtension.cmake
@@ -106,4 +106,17 @@ macro (copy_python_in_build)
         file(COPY ${py_file_name} DESTINATION ${CMAKE_CURRENT_BINARY_DIR}/${parents})
     endforeach()
 
+    # Copy package data needed by Python modules at runtime as well. This is
+    # required for in-tree builds, where the install(DIRECTORY ...) rule below
+    # has not yet run.
+    file(GLOB_RECURSE package_data_files
+        "${TUDATPY_SOURCE_DIR}/*.csv"
+        "${TUDATPY_SOURCE_DIR}/*.json"
+    )
+    foreach(package_data_file ${package_data_files})
+        file(RELATIVE_PATH package_data_file_name ${TUDATPY_SOURCE_DIR} ${package_data_file})
+        get_filename_component(parents ${package_data_file_name} DIRECTORY)
+        file(COPY ${package_data_file_name} DESTINATION ${CMAKE_CURRENT_BINARY_DIR}/${parents})
+    endforeach()
+
 endmacro()
